gifts smallint to int, PG config, gift controller error, gift config and helper

+++ app/Config/Schema/gifts.sql
@@ -3,7 +3,7 @@
 -- http://www.phpmyadmin.net
 --
 -- Host: localhost
--- Generation Time: Jan 16, 2012 at 03:57 PM
+-- Generation Time: Jan 20, 2012 at 02:39 PM
 -- Server version: 5.1.49
 -- PHP Version: 5.3.3-7+squeeze3
 
@@ -28,7 +28,7 @@ SET SQL_MODE="NO_AUTO_VALUE_ON_ZERO";
 CREATE TABLE IF NOT EXISTS `gifts` (
   `id` char(36) NOT NULL,
   `serial` mediumint(8) unsigned NOT NULL AUTO_INCREMENT,
-  `amount` smallint(6) unsigned NOT NULL,
+  `amount` int(10) unsigned NOT NULL,
   `currency` char(3) NOT NULL DEFAULT 'INR' COMMENT 'ISO 4217',
   `status` enum('pending','success','failure') NOT NULL DEFAULT 'pending',
   `statuscode` varchar(16) DEFAULT NULL,
